Title: 학점 계산기, Time: 6ms, Memory: 0MB, Status: Passed - CodeTree

+++ 231230/학점 계산기/credit-calculator.c
@@ -4,13 +4,13 @@ int main() {
     // 여기에 코드를 작성해주세요.
     int n=0;
     scanf("%d",&n);
-    float a[10];
-    float sum=0;int i;
+    double a[10];
+    double sum=0;int i;
     for( i=0;i<n;i++){
-        scanf("%f",&a[i]);
+        scanf("%lf",&a[i]);
         sum+=a[i];
     }
-    printf("%.1f\n",sum/n);
+    printf("%.1lf\n",sum/n);
     if((float)sum/n>=4){
         printf("Perfect");
